# api/src/main/java/ee/biceps/api/generated/api/BusLinesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ee.biceps.api.generated.api;

import ee.biceps.api.generated.model.BusLine;
import ee.biceps.api.generated.model.ErrorMessage;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-01-30T23:16:18.141542013+02:00[Europe/Tallinn]")
@Validated
@Tag(name = "bus-lines", description = "the bus-lines API")
public interface BusLinesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /bus-lines : Create a bus line
     * Creates a bus line with given name and bus stops.
     *
     * @param busLine A new bus line object with name and list of bus stops. (optional)
     * @return Returned a bus line with fullfilled id field (status code 201)
     *         or One of the given bus stops do not exist. (status code 400)
     */
    @Operation(
        operationId = "createBusLine",
        summary = "Create a bus line",
        tags = { "LineMangagment" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Returned a bus line with fullfilled id field", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BusLine.class))
            }),
            @ApiResponse(responseCode = "400", description = "One of the given bus stops do not exist.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorMessage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/bus-lines",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<BusLine> _createBusLine(
        @Parameter(name = "BusLine", description = "A new bus line object with name and list of bus stops.") @Valid @RequestBody(required = false) BusLine busLine
    ) {
        return createBusLine(busLine);
    }

    // Override this method
    default  ResponseEntity<BusLine> createBusLine(BusLine busLine) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"busStopIds\" : [ 0, 0 ], \"name\" : \"Kummeli - Ringtee\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /bus-lines/{busLineId} : Deletes a bus-line
     * Searches for a specific bus-line with given id and deletes it from database.
     *
     * @param busLineId Common ID parameter of bus-line. (required)
     * @return Bus-line with given ID deleted successfully. (status code 204)
     *         or Bus-line with given ID is active and cannot be deleted. (status code 400)
     *         or Bus-line with given ID not found error. (status code 404)
     */
    @Operation(
        operationId = "deleteBusLine",
        summary = "Deletes a bus-line",
        responses = {
            @ApiResponse(responseCode = "204", description = "Bus-line with given ID deleted successfully."),
            @ApiResponse(responseCode = "400", description = "Bus-line with given ID is active and cannot be deleted.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorMessage.class))
            }),
            @ApiResponse(responseCode = "404", description = "Bus-line with given ID not found error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorMessage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/bus-lines/{busLineId}",
        produces = { "application/json" }
    )
    default ResponseEntity<Void> _deleteBusLine(
        @Parameter(name = "busLineId", description = "Common ID parameter of bus-line.", required = true) @PathVariable("busLineId") Integer busLineId
    ) {
        return deleteBusLine(busLineId);
    }

    // Override this method
    default  ResponseEntity<Void> deleteBusLine(Integer busLineId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /bus-lines : Get all bus lines with given filters.
     * Queries all bus lines with given optional name, busStopId and limit filters.
     *
     * @param maxResults The numbers of items to return. (optional, default to 20)
     * @param name Optional parameter to filter bus lines by partial name. (optional)
     * @param busStopId Optional parameter to filter bus lines by bus stops. (optional)
     * @return Returned a list of bus stops which fullfill query params. (status code 200)
     */
    @Operation(
        operationId = "readBusLines",
        summary = "Get all bus lines with given filters.",
        tags = { "LineMangagment" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Returned a list of bus stops which fullfill query params.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BusLine.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/bus-lines",
        produces = { "application/json" }
    )
    default ResponseEntity<List<BusLine>> _readBusLines(
        @Min(1) @Max(50) @Parameter(name = "maxResults", description = "The numbers of items to return.") @Valid @RequestParam(value = "maxResults", required = false, defaultValue = "20") Integer maxResults,
        @Parameter(name = "name", description = "Optional parameter to filter bus lines by partial name.") @Valid @RequestParam(value = "name", required = false) String name,
        @Parameter(name = "busStopId", description = "Optional parameter to filter bus lines by bus stops.") @Valid @RequestParam(value = "busStopId", required = false) Integer busStopId
    ) {
        return readBusLines(maxResults, name, busStopId);
    }

    // Override this method
    default  ResponseEntity<List<BusLine>> readBusLines(Integer maxResults, String name, Integer busStopId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"busStopIds\" : [ 0, 0 ], \"name\" : \"Kummeli - Ringtee\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}